pyproject.toml
[tool.poetry]
name = "osf-io"
version = "24.05.0"
description = "The code for [https://osf.io](https://osf.io)."
authors = ["Your Name <you@example.com>"]
license = "Apache License 2.0"
readme = "README.md"
packages = [{include = "osf"}]

[tool.poetry.dependencies]
python = "^3.12"
wheel = "0.42.0"
invoke = "2.2.0"
Werkzeug = "3.0.1"
Flask = "3.0.2"
Mako = "1.3.2"
Markdown = "3.5.2"
WTForms = "3.1.2"
celery = "5.3.6"
kombu = "5.3.5"
itsdangerous = "2.1.2"
lxml = "5.1.0"
mailchimp3 = "3.0.21"
nameparser = "1.1.3"
bcrypt = "4.1.2"
python-dateutil = "2.9.0"
pytz = "2024.1"
bleach = {version= "6.1.0", extras = ["css"] }
pillow = "10.2.0"
Markupsafe = "2.1.5"
blinker = "1.7.0"
furl = "2.1.3"
elasticsearch2 = "2.5.1"
elasticsearch = "6.8.2" # max version to support elasticsearch6
elasticsearch-dsl = "6.4.0" # max version to support elasticsearch6
elastic-transport = "8.13.0"
google-api-python-client = "2.123.0"
google-auth = "2.29.0"
Babel = "2.14.0"
citeproc-py = "0.6.0"
boto3 = "1.34.60"
django-waffle = "4.1.0"
# oscp is required for pymongo to work with mongodb <=3.4 on python >3.10
pymongo ={version= "3.13.0", extras = ["ocsp"]} # install to get bson module
PyYAML = "6.0.1"
tqdm = "4.66.2"
email-validator = "2.1.1"
#" Python markdown extensions for comment emails
markdown-del-ins = "1.0.0"
certifi = "2024.2.2"
sendgrid = "6.11.0"
requests = "2.31.0"
urllib3 = "1.26.18" # still <2.0 because elasticseach2 lib doesn't supprort urllib3>=2.0
oauthlib = "3.2.2"
requests-oauthlib = "1.3.1"
sentry-sdk = {version= "2.2.0", extras = ["django", "flask", "celery"]}
django-redis = "5.4.0"
# API requirements
Django = "4.2.13"
djangorestframework = "3.15.1"
django-cors-headers = "4.3.1"
djangorestframework-bulk = "0.2.1"
django-bulk-update = "2.2.0"
hashids = "1.3.1"
pyjwt = "2.8.0"
django-celery-beat = "2.6.0"
django-celery-results = "2.5.1"
django-timezone-field = "6.1.0" # pin to delay untested major version bump
pyjwe = "1.0.0"
# Required by pyjwe and ndg-httpsclient
cryptography = "42.0.5"
#rpds-py==0.18.0
jsonschema = "4.21.1"
django-guardian = "2.4.0"
# Admin requirements
django-webpack-loader = {git = "https://github.com/felliott/django-webpack-loader.git", rev = "af8438c2da909ec9f2188a6c07c9d2caad0f7e93"}
django-sendgrid-v5 = "1.2.3" # metadata says python 3.10 not supported, but tests pass

# Analytics requirements
keen = "0.7.0"
geoip2 = "4.7.0"
# OSF models
django-typed-models = "0.14.0"
django-storages = "1.14.3"
google-cloud-storage = "2.16.0"  # dependency of django-storages, hard-pin to version
django-dirtyfields = "1.9.2"
django-extensions = "3.2.3"
psycopg2 = "2.9.9"
# Reviews requirements
transitions = "0.8.11"
# identifiers
datacite = "1.1.3"
# metadata
rdflib = "7.0.0"
colorlog = "6.8.2"
# Metrics
django-elasticsearch-metrics = {git ="https://github.com/CenterForOpenScience/django-elasticsearch-metrics", branch = "feature/pin-esdsl"}
# Impact Metrics CSV Export
djangorestframework-csv = "3.0.2"
gevent = "24.2.1"
packaging = "^24.0"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.4"
pytest-socket = "0.7.0"
pytest-xdist = "3.5.0"
pytest-django = "4.8.0"
python-coveralls = "2.9.3" # do we need it?
pytest-testmon = "2.1.0"
pytest-asyncio = "0.23.5"
pytest-html = "4.1.1"
factory-boy = "3.3.0"
webtest-plus = "1.0.0"
Faker = "23.2.1"
schema = "0.7.4"
responses = "0.25.0"

# Syntax checking
flake8 = "7.0.0"
flake8-mutable = "1.2.0"
pre-commit = "3.7.1"

# Django Debug Toolbar for local development
django-debug-toolbar = "4.3.0"

# Ipdb
ipdb = "0.13.13"

# PyDevD (Remote Debugging)
pydevd = "3.0.3"

# n+1 query detection
nplusone = "1.0.0"

# Profiling
django-silk = "5.1.0"

[tool.poetry.group.addons.dependencies]
# Requirements for the boa add-on
boa-api = "0.1.14"

# Requirements for running asyncio in celery, using 3.4.1 for Python 3.6 compatibility
asgiref = "3.7.2"
boxsdk = "3.9.2"
# Allow for optional timeout parameter.
# https://github.com/IQSS/dataverse-client-python/pull/27
dataverse = {git = "https://github.com/CenterForOpenScience/dataverse-client-python.git", rev="2b3827578048e6df3818f82381c7ea9a2395e526"}
dropbox = "11.36.2"

cachecontrol = "0.14.0"
"github3.py" = "4.0.1"
uritemplate = "4.1.1"
python-gitlab = "4.4.0"
# up-to-date with mendeley's master + add folder support and future dep updates
mendeley = {git = "https://github.com/CenterForOpenScience/mendeley-python-sdk.git", rev="feature/osf-dep-updates"}
# Requirements for the owncloud add-on
pyocclient = "0.6.0"
boto3 = "1.34.60"
pyotp = "2.9.0"
Pyzotero = "1.5.18"

[tool.poetry.group.release.dependencies]
# newrelic APM agent
newrelic = "9.7.1"
# uwsgi
uwsgi = "2.0.24"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
